Create CapitalIQ transcript.py
Able to automate the download of transcripts after being logged into the research page of Capital IQ. The code will download all transcripts that have an earnings call marker on them. In order to utilize properly, I recommend using the filters to make sure they are set to North America and the years you desire to download.